Fix a bug in meal lookup
Processing the result of the AJAX request for all meals was dependent on meal ids being in order, which was not the case due to the asynchronous nature of the seeding process which is using Promise.all(). This commit changes meal lookup to use the name instead of the id, which should always result in a match.